refactor: extract config management into module
