+++ curation_gui.py
@@ -63,7 +63,7 @@ class CurationMainWindow(QMainWindow):
         if config_manager:
             self.config = config_manager
         else:
-            from rom_curator_main import ConfigManager
+            from config_manager import ConfigManager
             self.config = ConfigManager()
         
         self.init_ui()
