+++ rom_curator_main.py
@@ -6,11 +6,8 @@ Provides a unified interface with menu options for all tools and workflows.
 """
 
 import sys
-import json
 import logging
-from copy import deepcopy
 from pathlib import Path
-from typing import Any, Dict, Literal
 from PyQt5.QtWidgets import (
     QApplication, QMainWindow, QWidget, QVBoxLayout, QHBoxLayout,
     QMenuBar, QStatusBar, QAction, QMessageBox, QLabel, QPushButton,
@@ -20,291 +17,13 @@ from PyQt5.QtCore import Qt, QTimer, pyqtSignal
 from PyQt5.QtGui import QPixmap, QFont, QIcon
 import qdarkstyle
 
-from pydantic import BaseModel, ConfigDict, Field, ValidationError, field_validator
-
-logger = logging.getLogger(__name__)
-
-
-def _default_ingestion_settings() -> Dict[str, Any]:
-    """Return default ingestion settings."""
-
-    return {
-        "library_roots": [],
-        "batch_size": 100,
-        "enable_validation": True,
-        "enable_archive_expansion": True,
-        "hash_algorithms": ["sha1", "crc32", "md5", "sha256"],
-        "file_extensions": {
-            "rom": [
-                ".rom",
-                ".bin",
-                ".smd",
-                ".sfc",
-                ".nes",
-                ".gb",
-                ".gba",
-                ".nds",
-                ".iso",
-                ".img",
-            ],
-            "archive": [".zip", ".7z", ".rar", ".tar", ".gz"],
-        },
-        "max_file_size_mb": 1024,
-        "exclude_patterns": ["*.tmp", "*.temp", "*.bak", "*.backup"],
-        "enable_platform_detection": True,
-        "enable_metadata_extraction": True,
-    }
-
-
-class GuiSettings(BaseModel):
-    """GUI specific configuration."""
-
-    model_config = ConfigDict(extra="allow")
-
-    window_width: int = Field(default=1200, ge=800, le=3840)
-    window_height: int = Field(default=800, ge=600, le=2160)
-    theme: Literal["dark", "light"] = "dark"
-
-
-class RomCuratorConfig(BaseModel):
-    """Root configuration model."""
-
-    model_config = ConfigDict(extra="allow")
-
-    auto_create_directories: bool = True
-    database_path: Path = Field(default=Path("./database/RomCurator.db"))
-    importer_scripts_directory: Path = Field(default=Path("./scripts/seeders/"))
-    log_directory: Path = Field(default=Path("./logs/"))
-    log_level: str = "INFO"
-    progress_update_interval: int = Field(default=100, ge=1, le=1000)
-    gui_settings: GuiSettings = Field(default_factory=GuiSettings)
-    ingestion_settings: Dict[str, Any] = Field(
-        default_factory=_default_ingestion_settings
-    )
-
-    @field_validator("log_level", mode="before")
-    @classmethod
-    def _validate_log_level(cls, value: Any) -> str:
-        level = str(value).upper()
-        if level not in {"DEBUG", "INFO", "WARNING", "ERROR"}:
-            raise ValueError("Log level must be one of DEBUG, INFO, WARNING, ERROR")
-        return level
-
-    @field_validator(
-        "database_path", "importer_scripts_directory", "log_directory", mode="before"
-    )
-    @classmethod
-    def _expand_path(cls, value: Any) -> Path:
-        if isinstance(value, Path):
-            return value.expanduser()
-        return Path(str(value)).expanduser()
-
-    @field_validator("database_path", mode="after")
-    @classmethod
-    def _validate_database_path(cls, value: Path, info) -> Path:
-        parent = value.parent
-        auto_create = info.data.get("auto_create_directories", True)
-        if not parent.exists():
-            if auto_create:
-                try:
-                    parent.mkdir(parents=True, exist_ok=True)
-                except OSError as exc:
-                    raise ValueError(
-                        f"Unable to create directory '{parent}': {exc}"
-                    ) from exc
-            else:
-                raise ValueError(
-                    f"Database directory '{parent}' does not exist and auto creation is disabled."
-                )
-        return value
-
-    @field_validator("importer_scripts_directory", "log_directory", mode="after")
-    @classmethod
-    def _validate_directories(cls, value: Path, info) -> Path:
-        auto_create = info.data.get("auto_create_directories", True)
-        if not value.exists():
-            if auto_create:
-                try:
-                    value.mkdir(parents=True, exist_ok=True)
-                except OSError as exc:
-                    raise ValueError(
-                        f"Unable to create directory '{value}': {exc}"
-                    ) from exc
-            else:
-                raise ValueError(
-                    f"Directory '{value}' does not exist and auto creation is disabled."
-                )
-        return value
-
-# Import our modules
+from config_manager import ConfigManager
 from enhanced_importer_gui import ImporterApp
 from curation_gui import CurationMainWindow
 from platform_linking_gui import PlatformLinkingDialog
 from extension_registry_gui import ExtensionRegistryDialog
-
-# Import enhanced logging system (Work Item 2)
 from enhanced_logging import EnhancedLoggingManager
-
-# Import resilient worker components (Work Item 2)
 from resilient_worker import ResilientWorkerThread
-
-
-class ConfigManager:
-    """Centralized configuration management using pydantic models."""
-
-    def __init__(self, config_file: str = 'config.json') -> None:
-        self.config_file = Path(config_file)
-        self._logger = logger
-        self._config_model = self._load_model()
-        self.config: Dict[str, Any] = self._config_model.model_dump(mode="json")
-
-    @property
-    def model(self) -> RomCuratorConfig:
-        """Return the typed configuration model."""
-
-        return self._config_model
-
-    def load_config(self) -> Dict[str, Any]:
-        """Reload configuration from disk."""
-
-        self._config_model = self._load_model()
-        self.config = self._config_model.model_dump(mode="json")
-        return self.config
-
-    def _load_model(self) -> RomCuratorConfig:
-        """Load and validate configuration, falling back to defaults on error."""
-
-        if not self.config_file.exists():
-            default_model = RomCuratorConfig()
-            self._write_config_file(default_model)
-            return default_model
-
-        try:
-            with open(self.config_file, 'r', encoding='utf-8') as handle:
-                raw_config = json.load(handle)
-        except (json.JSONDecodeError, OSError) as exc:
-            self._logger.error("Failed to read configuration: %s", exc)
-            self._show_error_dialog(
-                "Configuration Error",
-                "The configuration file could not be read. Default settings will be used.",
-            )
-            default_model = RomCuratorConfig()
-            self._write_config_file(default_model)
-            return default_model
-
-        try:
-            return RomCuratorConfig.model_validate(raw_config)
-        except ValidationError as error:
-            self._logger.error("Configuration validation failed: %s", error)
-            self._show_error_dialog("Invalid Configuration", self._format_validation_errors(error))
-            default_model = RomCuratorConfig()
-            self._write_config_file(default_model)
-            return default_model
-
-    def get(self, key: str, default: Any = None) -> Any:
-        """Get a configuration value by dotted path."""
-
-        keys = key.split('.')
-        value: Any = self.config
-
-        for part in keys:
-            if isinstance(value, dict) and part in value:
-                value = value[part]
-            else:
-                return default
-
-        return value
-
-    def set(self, key: str, value: Any) -> None:
-        """Set a configuration value and revalidate the model."""
-
-        keys = key.split('.')
-        updated_config = deepcopy(self.config)
-        ref = updated_config
-
-        for part in keys[:-1]:
-            if part not in ref or not isinstance(ref[part], dict):
-                ref[part] = {}
-            ref = ref[part]
-
-        ref[keys[-1]] = value
-
-        try:
-            self._config_model = RomCuratorConfig.model_validate(updated_config)
-        except ValidationError as error:
-            self._logger.error("Invalid configuration update for '%s': %s", key, error)
-            self._show_error_dialog("Invalid Configuration", self._format_validation_errors(error))
-            return
-
-        self.config = self._config_model.model_dump(mode="json")
-
-    def save(self) -> None:
-        """Persist configuration to disk after validation."""
-
-        try:
-            self._config_model = RomCuratorConfig.model_validate(self.config)
-        except ValidationError as error:
-            self._logger.error("Configuration invalid during save: %s", error)
-            self._show_error_dialog("Invalid Configuration", self._format_validation_errors(error))
-            return
-
-        self.config = self._config_model.model_dump(mode="json")
-        self._write_config_file(self._config_model)
-
-    def ensure_directories(self) -> None:
-        """Create required directories if they do not exist."""
-
-        if not self.get('auto_create_directories', True):
-            return
-
-        directories = [
-            self._config_model.database_path.parent,
-            self._config_model.log_directory,
-            self._config_model.importer_scripts_directory,
-        ]
-
-        for directory in directories:
-            directory.mkdir(parents=True, exist_ok=True)
-
-    def _write_config_file(self, config_model: RomCuratorConfig) -> None:
-        """Write the provided configuration model to disk."""
-
-        data = config_model.model_dump(mode="json")
-        try:
-            parent = self.config_file.parent
-            if parent and not parent.exists():
-                parent.mkdir(parents=True, exist_ok=True)
-            with open(self.config_file, 'w', encoding='utf-8') as handle:
-                json.dump(data, handle, indent=4)
-        except OSError as exc:
-            self._logger.error("Could not save config file: %s", exc)
-
-    @staticmethod
-    def _show_error_dialog(title: str, message: str) -> None:
-        """Display an error dialog if a QApplication is active."""
-
-        app = QApplication.instance()
-        if app is not None:
-            QMessageBox.critical(None, title, message)
-
-    @staticmethod
-    def _format_validation_errors(error: ValidationError) -> str:
-        """Format validation errors for display."""
-
-        error_lines = [
-            "The configuration file contains invalid values. Default settings will be used.",
-            "",
-        ]
-
-        for detail in error.errors():
-            location = " > ".join(str(part) for part in detail.get('loc', ()))
-            message = detail.get('msg', 'Invalid value')
-            if location:
-                error_lines.append(f"- {location}: {message}")
-            else:
-                error_lines.append(f"- {message}")
-
-        return "\n".join(error_lines)
 
 
 class LoggingManager:
